Refactor WorldFormulaEngine to use UniversalId instead of DID
- Updated contributions HashMap to use UniversalId as the key.
- Adjusted methods to handle UniversalId for contributions, including insertion and retrieval.
- Changed temporal weight calculation to use Lamport timestamps instead of DateTime.
- Modified trust normalization to ensure consistency with f64 type.
- Updated tests to reflect changes in UniversalId usage.

Refactor WorldFormulaContribution to remove legacy builder API

- Removed legacy builder methods for creating contributions.
- Updated constructors to use the modern API.
- Ensured trust normalization is calculated correctly upon initialization.

Update ErynoaHost to support LegacyDID for compatibility

- Adjusted DID parsing to use LegacyDID for trust and identity store compatibility.

Refactor ManaAccount to use f32 for trust snapshot

- Changed trust_snapshot type to f32 for compatibility with unified TrustVector6D.

Update RealmStorage to use UniversalId for owner_did

- Changed owner_did to store the hex representation of UniversalId.

Refactor GatewayGuard to use UniversalId for trust vectors and credentials

- Updated HashMaps to use UniversalId instead of DID for trust vectors and credentials.
- Adjusted methods to handle UniversalId for trust and credential registration.

Refactor IntentParser to use UniversalId for intents

- Updated intent creation to use UniversalId for subject and target.
- Adjusted timeout handling to use seconds instead of Duration.

Refactor SagaComposer to use UniversalId for intents

- Updated methods to handle UniversalId for intent creation and validation.

Refactor AnomalyDetector to use UniversalId for event history

- Changed event history and transfer graph to use UniversalId instead of DID.
- Updated anomaly detection logic to work with Lamport timestamps.

+++ backend/src/domain/formula.rs
@@ -228,7 +228,7 @@ impl WorldFormulaContribution {
 
         // ‖𝕎(s)‖_w
         let weights = self.context.weights();
-        let trust_norm = self.trust.weighted_norm(&weights);
+        let trust_norm = self.trust.weighted_norm(&weights) as f64;
 
         // ln|ℂ(s)|
         let log_history = (self.causal_history_size.max(1) as f64).ln();
